pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "Cython"]
build-backend = "setuptools.build_meta"

[project]
name = "macworp"
version = "2.0.0"
requires-python = ">=3.12"
license = {file = "./LICENSE"}
description = "Shared code for Massive Accessible Workflow Platform (MAcWorP) backend and worker"
authors = [
    {name = "Core Unit Bioinformatic Ruhr University Bochum", email = "cubimed@rub.de"},
]
maintainers = [
    {name = "Dirk Winkelhardt", email = "dirk.winkelhardt@rub.de"}
]

dependencies = [
	"GitPython ~=3.1",
	"fastapi >= 0.110.2",
	"jsonschema",
	"mergedeep >=1.3.4, <2",
	"nicegui",
	"oauthlib",
	"pandas",
	"parameterized ~= 0.9",
	"passlib ~= 1.7",
	"pika",
	"psycopg ~= 3.1",
	"pydantic-yaml",
	"pyjwt ~= 2.8",
	"python-jose[cryptography]",
	"python-multipart",
	"pyyaml ~= 6.0",
	"requests >=2, <3",
	"sqlmodel >= 0.0.24",
	"uvicorn[standard] >= 0.29.0",
	"snakemake ~= 9.5",
	"typing_extensions",
]

[project.optional-dependencies]
dev = [
	"black",
    "honcho",
	"pandas-stubs",
    "pylint",
	"sphinx",
	"types-PyYAML",
]

# Dependencies for the demo Workflows
demo = [
    "kaleido ~= 1.1.0",
    "plotly ~= 6.3.0"
]

[project.scripts]
macworp = "macworp.__main__:main"

[tool.setuptools]
include-package-data = true